thaidiy/requirements.txt
boto3==1.9.96
botocore==1.12.96
astroid==2.2.5
autopep8==1.4.3
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.7
git+git://github.com/MatthewGaber/django-ckeditor.git
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
django-js-asset==1.2.2
dnspython==1.16.0
docopt==0.6.2
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
monotonic==1.5
numpy==1.16.2
pep8==1.7.1
Pillow==6.0.0
pipreqs==0.4.9
psycopg2==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pyparsing==2.3.1
python-dateutil==2.8.0
python-engineio==3.4.3
python-socketio==3.1.2
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.2.18
typed-ast==1.3.4
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
WTForms==2.2.1
yarg==0.1.9
